requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
scipy>=1.7.0
annoy>=1.17.0
rapidfuzz>=2.0.0
pytest>=7.0.0
python-telegram-bot[job-queue]>=20.0
python-dotenv>=1.0.0
langdetect>=1.0.9
jupyter>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0 